Don't delete the original if we decided to use the original unscaled for the thumbnail image.

+++ src/java/com/threerings/msoy/web/server/UploadServlet.java
@@ -178,7 +178,7 @@ public class UploadServlet extends HttpServlet
 
         // if the user is uploading a thumbnail image, we want to use the scaled version and
         // abandon their original
-        if (tinfo != null && mediaId.equals(Item.THUMB_MEDIA)) {
+        if (tinfo != null && !tinfo.hash.equals(info.hash) && mediaId.equals(Item.THUMB_MEDIA)) {
             info = tinfo;
             tinfo = null;
             if (!output.delete()) {
